Expunge AI artwork per The Unwritten Vukky Rules, Section 727, Clause 42
Every time I see these pop up in #dumpster, I feel embarrassed for some reason or another. I suspect this is for multiple reasons. If you're the-person-who-can't-decide-on-a-name, I'm sure I already told you these things, but I'll write it here in longer form for those who are not that person.

I'm not much of an artist myself, but I do understand it takes quite an amount of effort to get to the point where you can make a masterpiece. I can see why you would think that effort would be wasted when the computer can print art faster than a human reasonably would. It makes me feel bad!

Another reason is that the art... just isn't that good? Take a look at 00037 for example. That's someone throwing a very vaguely Sun-looking object, but it's certainly not Niko. Reminds me more of Klee from Genshin Impact. I'm sure you could think of a better example, but I can't. 00095 is my favorite one of the bunch, but there's something off about the colors.

If you were to just commission someone instead, I wouldn't be surprised if they could figure out what you wanted in one go. Sure, it costs money, but there's nothing wrong with paying for something... right?

Now, you might be disagreeing with something I'm saying here. Maybe I'll disagree with it in the future, or see something wrong about it. Too bad! I wrote it in one go, it's eternal now.

Ok bye òwó